refactor: view 관련 코드는 domain 모듈이 아닌 app 모듈에 위치 시키기
