+++ app/src/main/java/com/github/dodobest/upbitapi/model/Coin.kt
@@ -1,4 +1,4 @@
-package com.github.dodobest.upbitapi
+package com.github.dodobest.upbitapi.model
 
 enum class Coin(val koreanName: String) {
     NO_EXIST("NO EXIST");
