+++ domain/src/main/java/com/github/dodobest/domain/model/UpbitTickerData.kt
@@ -6,11 +6,4 @@ data class UpbitTickerData(
     val tradePrice: Double,
     val signedChangeRate: Double,
     val aacTradePrice24h: Double,
-) {
-
-    fun addKoreanName(koreanName: String): UpbitTickerDataWithKoreanName {
-        return  UpbitTickerDataWithKoreanName(
-            market, koreanName, openingPrice, tradePrice, signedChangeRate, aacTradePrice24h
-        )
-    }
-}
+)
